Refactor the mocks a little.
